Moved logic for enum/struct derive into their own modules.
